♻️ refactor: 이미지 상세보기 훅으로 분리
