+++ src/components/Feed/FeedCard/FeedCard.module.scss
@@ -95,6 +95,7 @@
     height: 212px;
     flex-shrink: 0;
     border-radius: 5px;
+    cursor: pointer;
   }
 }
