+++ src/components/Post/Markdown/index.tsx
@@ -1,13 +1,9 @@
-import Markdown from 'react-markdown';
-import remarkGfm from 'remark-gfm';
-import rehypeRaw from 'rehype-raw';
-import CodeBlock from './CodeBlock';
-import classNames from 'classnames/bind';
-import styles from './Markdown.module.scss';
-import Modal from '@/components/Common/Layout/Modal';
+import DetailImageModal from '@/components/Common/DetailImageModal';
 import { useImageDetail } from '@/hooks/useImageDetail';
-
-const cn = classNames.bind(styles);
+import Markdown from 'react-markdown';
+import rehypeRaw from 'rehype-raw';
+import remarkGfm from 'remark-gfm';
+import CodeBlock from './CodeBlock';
 
 interface MarkdownContentProps {
   className: string;
@@ -51,23 +47,11 @@ export default function MarkdownContent({
       >
         {content}
       </Markdown>
-      <Modal
-        modalVisible={isImageModalVisible}
-        cssComponentDisplay={cn('modal-container')}
-        cssModalSize={cn('modal-wrapper')}
-        toggleModal={hideImageDetail}
-      >
-        <img
-          src={currentImageUrl}
-          alt="detail"
-          style={{
-            objectFit: 'contain',
-            width: '800px',
-            maxHeight: '80vh',
-            maxWidth: '80vw',
-          }}
-        />
-      </Modal>
+      <DetailImageModal
+        currentImageUrl={currentImageUrl}
+        isImageModalVisible={isImageModalVisible}
+        hideImageDetail={hideImageDetail}
+      />
     </>
   );
 }
